Revamped makefile
Apparently it is not good to override CXX in a makefile,
so I have made that an option. This required different
flags depending on which compiler is being used. Also,
while I was messing with the makefile, I decided to try
enabling 'make release' instead of 'make CONFIG=Release'
This means the object files cannot be differentiated
based on build type, and must be rebuilt when a different
build type was used last.