tblyons/cocles#4: fixed
Ledger::payees() would query the extended adjustments sequence for
the payees. Since the database is now using a separate sorted,
unique sequence to track Payees, it should search that list for the
Payees that match the requested regex.

As a result of identifying this fix, it became apparent that
Ledger::accounts() works in the same broken manner. I will create
a new issue and attempt to tease a test case that will cause
accounts to be reported multiple times.