tblyons/cocles#5: fixed
Ledger::accounts() was updated to query the separate, sorted, unique
sequence tracking accounts instead of the sequence of adjustments.

Right now, account objects are duplicated across m_accounts and
m_adjustments. I think this should probably be fixed, but I'm
not sure of a use case where it is necessary.